Fix small bug in the wrapper.
